+++ changelog.md
@@ -1,5 +1,6 @@
 ## 1.1
 * Fixed a grammatical error with bulls and cows plural output
+* Removed debug output that gave out the guessed number so game makes sense now
 
 ## Version 1.0
 * Initial release