+++ changelog.md
@@ -1,2 +1,9 @@
+## Version 1.1
+* Fixed a bug with digits repetition in guessed number
+* Added check for length of input number
+* Fixed a grammatical error with bulls and cows plural output
+* Removed debug output that gave out the guessed number so game makes sense now
+* Fixed a small bug that caused printing 4 bulls when you won no matter how many there actually is
+
 ## Version 1.0
 * Initial release